Implementa LogViewer,Health.
Corrige DeleteTemporaryFileJob.
Adiciona Horizon e schedule ao Docker.
Remove pacotes e arquivos não usados.

+++ app/Jobs/Tool/DeleteTemporaryFileJob.php
@@ -9,6 +9,7 @@ use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Queue\SerializesModels;
 use Illuminate\Support\Carbon;
 use Illuminate\Support\Facades\File;
+use Symfony\Component\Finder\Exception\DirectoryNotFoundException;
 
 class DeleteTemporaryFileJob implements ShouldQueue
 {
@@ -17,28 +18,33 @@ class DeleteTemporaryFileJob implements ShouldQueue
     /**
      * @var string
      */
-    protected $path;
+    public $path;
 
-    /**
-     *
-     */
-    public function __construct()
+    public function __construct($path = null)
     {
-        $this->path = storage_path('app/tmp/');
+        $this->path = $path ?? storage_path('app/tmp/');
     }
 
     /**
      * Deleta arquivos cujo o tempo tenha passado mais que 1h
-     * @return void
      */
     public function handle()
     {
-        foreach (File::files($this->path) as $file) {
-            if ($file->isFile()) {
-                if (now()->gt(Carbon::parse($file->getATime())->addHour())) {
-                    File::delete([$file->getPathname()]);
+        try {
+            if (File::isDirectory($this->path)) {
+                foreach (File::files($this->path) as $file) {
+                    if ($file->isFile()) {
+                        if (now()->gt(Carbon::parse($file->getATime())->addHour())) {
+                            File::delete([$file->getPathname()]);
+                        }
+                    }
                 }
+            } else {
+                File::makeDirectory($this->path);
             }
+        } catch (\Throwable $throwable) {
+            report($throwable);
         }
+
     }
 }
